Add support for reading correlators from DB
The previous I/O solution envisioned always
caching correlators into and hdf5 file and then
reading from there. Due to file permissions
(and the fact that appending to existing hdf5
datasets can be annoying), it's also desirable to
be able to read a correlator quickly from SQLite
and apply the tsm with a single function.
The present commit adds this functionality.